Develop staging
* added getResultsFromCalendarIncidence function
* limiting upload tests to windows os
* Update test-againstCDM.R
* Refactored code in index event breakdown.
* Consistently use local function renderTranslateQuerySql
* cdmDatasource information
* documentation for time series, updated time series

+++ R/CohortOverlapDiagnostics.R
@@ -16,79 +16,71 @@
 
 
 
-#' Given two sets of cohorts get relationships between the cohorts.
+#' Given two sets of cohorts get overlap between the cohorts.
 #'
 #' @description
-#' Given two cohorts, get data to compare relationships between the cohorts - such as overlap, temporal 
-#' relationships between cohort start date(s). 
+#' Given two sets of cohorts, get data on overlap between the cohorts.
 #' Note: only the first occurrence of subject_id in the cohort is used.
 #'
 #' @template Connection
-#' 
+#'
 #' @template CohortDatabaseSchema
 #'
 #' @template CohortTable
-#'                                    
-#' @param targetCohortIds             A list of cohort ids to be used as target cohorts.
-#' 
-#' @param comparatorCohortIds         A list of cohort ids to be used as comparator cohorts.
-#'                                    
-#' @param batchSize                   {Optional, default set to 200} If running diagnostics on larget set
-#'                                    of cohorts, this function allows you to batch them into chunks that run 
-#'                                    as a batch.
-#'                                    
+#'
+#' @template CohortIds
+#'
+#' @param batchSize                   {Optional, default set to 50} If running diagnostics on large set
+#'                                    of cohorts, this function allows you to batch them into chunks that
+#'                                    by default run over 50 target cohorts (and all comparator cohorts).
+#'
 #' @export
 computeCohortOverlap <- function(connectionDetails = NULL,
                                  connection = NULL,
                                  cohortDatabaseSchema,
                                  cohortTable = "cohort",
-                                 targetCohortIds,
-                                 comparatorCohortIds,
-                                 batchSize = 200) {
+                                 cohortIds,
+                                 batchSize = 50) {
   startTime <- Sys.time()
   
+  if (length(cohortIds) == 0) {
+    return(NULL)
+  }
   if (is.null(connection)) {
     connection <- DatabaseConnector::connect(connectionDetails)
     on.exit(DatabaseConnector::disconnect(connection))
   }
-  
-  cohortIds <- c(targetCohortIds, comparatorCohortIds) %>% unique() %>% sort()
   
   results <- Andromeda::andromeda()
   for (start in seq(1, length(cohortIds), by = batchSize)) {
     end <- min(start + batchSize - 1, length(cohortIds))
     if (length(cohortIds) > batchSize) {
       ParallelLogger::logInfo(sprintf(
-        "Batch characterization. Processing cohorts %s through %s",
+        "Batch Cohort Overlap Processing cohorts %s through %s",
         start,
         end
       ))
     }
-    
     sql <- SqlRender::loadRenderTranslateSql(
       "CohortOverlap.sql",
       packageName = "CohortDiagnostics",
       dbms = connection@dbms,
       cohort_database_schema = cohortDatabaseSchema,
       cohort_table = cohortTable,
-      target_cohort_ids = targetCohortIds,
-      comparator_cohort_ids = comparatorCohortIds
+      target_cohort_ids = cohortIds
     )
-    DatabaseConnector::executeSql(
+    DatabaseConnector::executeSql(connection = connection,
+                                  sql = sql)
+    
+    overlap <- renderTranslateQuerySql(
       connection = connection,
-      sql = sql
+      sql = "SELECT * FROM #cohort_overlap_long;",
+      snakeCaseToCamelCase = TRUE
     )
     
-    overlap <- DatabaseConnector::renderTranslateQuerySql(connection = connection, 
-                                                          sql = "SELECT * FROM #cohort_overlap_long;", 
-                                                          snakeCaseToCamelCase = TRUE)
-    
-    DatabaseConnector::renderTranslateExecuteSql(
-      connection = connection,
-      sql = "IF OBJECT_ID('tempdb..#cohort_overlap_long', 'U') IS NOT NULL DROP TABLE #cohort_overlap_long;",
-      progressBar = TRUE
-    )
-    
+    DatabaseConnector::renderTranslateExecuteSql(connection = connection,
+                                                 sql = "IF OBJECT_ID('tempdb..#cohort_overlap_long', 'U') IS NOT NULL DROP TABLE #cohort_overlap_long;",
+                                                 progressBar = TRUE)
     
     if ("overlap" %in% names(results)) {
       Andromeda::appendToTable(results$overlap, overlap)
